requirements.txt
beautifulsoup4==4.12.3
cairocffi==1.7.0
CairoSVG==2.7.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cssselect2==0.7.0
defusedxml==0.7.1
idna==3.7
lxml==5.2.2
neo4j==5.20.0
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pycairo==1.26.0
pycparser==2.22
pymediawiki==0.7.4
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.0
requests==2.32.3
six==1.16.0
soupsieve==2.5
svglib==1.5.1
tinycss2==1.3.0
tzdata==2024.1
urllib3==2.2.1
Wand==0.6.13
webencodings==0.5.1
